Move ConfirmationModal to shared components
